requirements.txt
google-generativeai>=0.4.0
PyPDF2>=3.0.0
python-docx>=0.8.11
scikit-learn>=1.0.0
numpy>=1.20.0



